+++ packages/uilib/src/lib/plugins/dedupe/dedupe.svelte
@@ -1,11 +1,12 @@
 
 
 <script lang="ts">
-	import { UCTypeDedupe, SCDQueries, type HashedDOT, type DOElement, type DOTypeElement, type SCDElement } from "@oscd-plugins/core"
+	import { UCTypeDedupe, SCDQueries, type SCDElement, type HashedTypeElement, type TypeElement } from "@oscd-plugins/core"
 	import GroupCardList from "./group-card-list/group-card-list.svelte"
 	import Merger from "./merger/merger.svelte"
 	import { NullParentElement, type MergableItem, type ParentElement } from "./merger/mergable-items"
 	import Theme from "../../style/theme.svelte"
+	import { ButtonGroup, type ButtonGroupOption } from "../../components/button-group"
 
 	// Input
 	export let doc: Element
@@ -24,20 +25,39 @@
 		if(!document){ return }
 		scdQueries  = new SCDQueries(document)
 		deduper = new UCTypeDedupe(scdQueries)
-		loadDuplicates()
 	}
 
-	let duplicateGroups: HashedDOT[][]
-	async function loadDuplicates(){
-		duplicateGroups = await deduper.findDuplicateObjectTypes()
+	
+	// $: loadDuplicates(selectedType)
+	let duplicateGroups: HashedTypeElement<TypeElement>[][] = []
+	$: loadDuplicates(selectedType)
+	async function loadDuplicates(type: string){
+		if(type === SCDQueries.SelectorDOType){
+			duplicateGroups =  await deduper.findDuplicateObjectTypes()
+			return 
+		}
+		if(type === SCDQueries.SelectorDAType) {
+			duplicateGroups = await deduper.findDuplicateDataAttributeTypes()
+			return
+		}
+		if(type === SCDQueries.SelectorLNodeType){
+			duplicateGroups = await deduper.findDuplicateLogicalNodeTypes()
+			return
+		}
+		if(type === SCDQueries.SelectorEnumType){
+			duplicateGroups = await deduper.findDuplicateEnumTypes()
+			return
+		}
+
+		duplicateGroups = []
+
 	}
 	
 	let itemSets: MergableItem[][] = []
-	$: convertDuplicatesToItems(duplicateGroups)
-	function convertDuplicatesToItems(groups?: HashedDOT[][]){
-		if(!groups){ return }
-
-		itemSets = groups.map((group) => {
+	$: itemSets = convertDuplicatesToItems(duplicateGroups)
+	function convertDuplicatesToItems(groups?: HashedTypeElement<TypeElement>[][]): MergableItem[][]{
+		if(!groups){ return []}
+		return groups.map((group) => {
 			return group.map((item) => {
 				return {
 					label:  item.element.id,
@@ -64,20 +84,56 @@
 	function handleGroupSelect(e: CustomEvent<{index:number}>){
 		selectedGroupIndex = e.detail.index
 	}
+
 	
+	const typeOptions: ButtonGroupOption[] = [
+		{ id: SCDQueries.SelectorLNodeType, label: "LNType" },
+		{ id: SCDQueries.SelectorDOType,    label: "DOType" },
+		{ id: SCDQueries.SelectorDAType,    label: "DAType" },
+		{ id: SCDQueries.SelectorEnumType,  label: "EnumType" },
+	]
+
+	let selectedType: string = typeOptions[0].id
+	function handleTypeChange(e: Event){
+		const target = e.target as HTMLInputElement
+		const type = target.value
+		selectedType = type
+	}
+
 
 	// 
 	// Merger
 	// 
 	let structure: string[] = []
 	$: loadStructureOfFirstElement(duplicateGroups, selectedGroupIndex)
-	function loadStructureOfFirstElement(groups: HashedDOT[][], index: number){
-		if(index < 0){ return }
+	function loadStructureOfFirstElement(groups: HashedTypeElement<TypeElement>[][], index: number){
+
+		const hasSelection = index > -1
+		if(!hasSelection){ 
+			resetStructure()
+			return
+		}
+
+		const hasGroups = groups.length > 0
+		if(!hasGroups){ 
+			resetStructure()
+			return	
+		}
+
 		const selectedGroup = groups[index]
+		const hasElements = selectedGroup?.length > 0
+		if(!hasElements){ 
+			resetStructure()
+			return	
+		}
+
 		const firstElement = selectedGroup[0]
 		const children = Array.from(firstElement.element.element.children)
 		const names = children.map((child) => child.getAttribute("name")??"")
 		structure = names.filter(Boolean)
+	}
+	function resetStructure(){
+		structure = []
 	}
 	
 
@@ -106,10 +162,10 @@
 		root.dispatchEvent(event)
 	}
 
-	function createRelinkActions(els: SCDElement, dot: DOTypeElement){
+	function createRelinkActions(els: SCDElement, typeEl: TypeElement){
 		const deep = true
 		const modifiedEl = els.element.cloneNode(deep) as Element
-		modifiedEl.setAttribute("type", dot.id)
+		modifiedEl.setAttribute("type", typeEl.id)
 
 		const actions = createEventDetail(els.element, modifiedEl)
 		return actions
@@ -138,6 +194,14 @@
 	<layout>
 		
 		<sidebar>
+			<h3>Duplicates</h3>
+			<div>
+				<ButtonGroup 
+					options={typeOptions} 
+					selectedID={typeOptions[0].id} 
+					on:change={handleTypeChange} 
+				/>
+			</div>
 			<GroupCardList 
 				itemSets={itemSets.map((itemSet) => itemSet.map((item) => item.label) )} 
 				on:select={handleGroupSelect} 
@@ -147,9 +211,9 @@
 		
 		<main>
 			{#if selectedGroupIndex > -1}
-			{#key selectedGroupIndex}
+			{#key `${selectedGroupIndex}_${selectedType}`}
 				<Merger 
-					items={itemSets[selectedGroupIndex]} 
+					items={itemSets[selectedGroupIndex]??[]} 
 					structure={structure}
 					on:merge={handleMerge}
 				/>
@@ -175,15 +239,18 @@
 	}
 	layout{
 		display: grid;
-		grid-template-columns: minmax(100px, 150px) 1fr;
+		grid-template-columns: 330px 1fr;
 		height: 100%;
 	}
 
 	sidebar{
-		border-right : black thin solid;
-		padding-right: 1rem;
-		height: 	   100%;
-		overflow: 	   auto;
+		border-right :  black thin solid;
+		padding-right:  1rem;
+		height: 	    100%;
+		overflow: 	    auto;
+		display:        flex;
+		flex-direction: column;
+		gap: 			1rem;
 	}
 
 </style>